feat(preview): Add new options win_width and win_vwidth
Similar to win_height and win_vheight, the newly introduced options
win_width and win_vwidth specifies the width of the preview window
(defaults remain the same, the full width).

The option values can be a real number between [0..1] (the ratio
relative to the full width) or a positive integer (characters).