fix(preview): gracefully handle nvim-treesitter main branch
Problem: The `main` branch of `nvim-treesitter` is not compatible with
the `master` branch. The API has changed:
- `nvim-treesitter.parsers` only contains a table with parser
  information, no additional functions
- `nvim-treesitter.configs` does not exist
- users should set up their own autocommands to start Treesitter in
  buffers, including the `qf` buffer.

When opening the quickfix list, nvim-bqf throws an error because it
cannot `require('nvim-treesitter.configs')`.

Solution: Detect when the user is on the `main` branch of
`nvim-treesitter`, and act as if Treesitter is not enabled at all.

If the user has an autocmd that calls `vim.treesitter.start()` in new
buffers, like `nvim-treesitter` recommends, then it will start
automatically for the preview buffer displayed by nvim-bqf.

This fixes the errors that are currently shown when opening the quickfix
list with `nvim-bqf` when using the `main` branch of `nvim-treesitter`.